Ossie #996: Moved buttons for execution out of AnalysisJobBuilderWindow
into a new class called ExecuteButtonBuilder
